requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
langchain
langchain-openai
langchain-pinecone
requests
pymupdf
pydantic>=2.0
structlog
cachetools
jinja2
openai
streamlit
fpdf2
python-multipart
semver
python-docx
datasets
tqdm
